split ops out
